Update patch set 1
Patch Set 1:

Uploaded as draft because I'm not sure this is the best way to solve it.

The root cause seems to be that the manifest is loaded more than once, but I suspect that fixing that may involve a much bigger change which may not be worth the effort just to fix such a trivial issue.

Patch-set: 1
Label: Code-Review=0